requirements.txt
discord.py
numpy
matplotlib
requests
colorama
dotenv
